% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_3kpk_score.R
\name{final_3kpk_score}
\alias{final_3kpk_score}
\title{Final 3KPK Score}
\usage{
final_3kpk_score(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Uses the scores & variables created during the prep scores call to find what should be the
final score proportions out of 5.
}
\details{
Should show up to 2 decimal places

To be used in a mutate in association with the prep scores call

INTERNAL FUNCTION FOR THIS PACKAGE
}
