% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_col_score.R
\name{final_col_score}
\alias{final_col_score}
\title{Final COL Score}
\usage{
final_col_score(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Calculates the final col score proportion out of 5
}
\details{
Should show up to 2 decimal places

To be used in a mutate within the prep_scores call. INTERNAL FUNCTION FOR THIS PACKAGE
}
