% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multiple_charters.R
\name{gen_multiple_charters}
\alias{gen_multiple_charters}
\title{gen_multiple_charters. Wrapper around generate_charter_export. This takes a list of dataframes and computes multiple exports at once}
\usage{
gen_multiple_charters(list_of_scores, template, path)
}
\arguments{
\item{list_of_scores}{Must be list of dataframes matching the expected output.}

\item{template}{excel template to use}

\item{path}{where to save files}
}
\value{

}
\description{
gen_multiple_charters. Wrapper around generate_charter_export. This takes a list of dataframes and computes multiple exports at once
}
