% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multiple_scores.R
\name{gen_multiple_scores}
\alias{gen_multiple_scores}
\title{Export a list of DataFrames to excel files.}
\usage{
gen_multiple_scores(list_of_scores, template, path)
}
\arguments{
\item{list_of_scores}{input list of dataframes}

\item{template}{excel template to write scores to}

\item{path}{file path for excels to be saved.}
}
\value{
Returns empty. Maps over a list of dataframes exporting them.
}
\description{
Export a list of DataFrames to excel files.
}
