% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_charter_export.R
\name{generate_charter_export}
\alias{generate_charter_export}
\title{Generate Charter Export. Only Generates one export. Expecting a single dataframe. Please use "prep_charter_scores"
to get the dataframe into the expected format}
\usage{
generate_charter_export(df, template, file_path_to_save)
}
\arguments{
\item{df}{a dataframe to enter}

\item{template}{excel template from passport to use}

\item{file_path_to_save}{file location to save to.}
}
\value{

}
\description{
Generate Charter Export. Only Generates one export. Expecting a single dataframe. Please use "prep_charter_scores"
to get the dataframe into the expected format
}
