% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_passport_exports.R
\name{generate_passport_exports}
\alias{generate_passport_exports}
\title{Generates excel file outputs for upload into passport for the 3kpk 1395 rfp}
\usage{
generate_passport_exports(df, template, file_path_to_save)
}
\arguments{
\item{df}{which dataset/readscores to use}

\item{template}{which excel template/export from passport to use}

\item{file_path_to_save}{where do you want the files saved. Character file path}
}
\value{
saves filled workbook to whichever directory you choose
}
\description{
Generates excel file outputs for upload into passport for the 3kpk 1395 rfp
}
\examples{


}
