% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_id.R
\name{generate_random_id}
\alias{generate_random_id}
\title{generate_random_id}
\usage{
generate_random_id(char_leng = 9)
}
\arguments{
\item{char_leng}{}
}
\value{

}
\description{
generate_random_id
}
