% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charter_datasets.R
\name{get_charter_datasets}
\alias{get_charter_datasets}
\title{Get necessary datasets for running the charter fse process}
\usage{
get_charter_datasets()
}
\value{

}
\description{
This will grab the necessary files & sheet info and save them as global variables. This also
keeps the nomenclature needed for using the sibling functions of prep_charter_scores and
generater_charter_export.
}
