% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nyc_zips.R
\name{get_nyc_zips}
\alias{get_nyc_zips}
\title{Get NYC Zip codes, borough, and neighborhood info from querying data.beta.nyc}
\usage{
get_nyc_zips()
}
\value{
A tibble with 6 variables: zip, borough, post_office, neighborhood, population, and density.
Data should read in as double/numeric for zip, pop, and density. Borough, post_office, and neighborhood should be character.
}
\description{
Get NYC Zip codes, borough, and neighborhood info from querying data.beta.nyc
}
\examples{
zips <- get_nyc_zips()
}
