% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rfp_datasets.R
\name{get_rfp_datasets}
\alias{get_rfp_datasets}
\title{Get RFP Datasets. This will read in all the necessary datasets for score exports for the 3kpk & col rfps.}
\usage{
get_rfp_datasets()
}
\value{

}
\description{
Get RFP Datasets. This will read in all the necessary datasets for score exports for the 3kpk & col rfps.
}
