% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in infix operator}
\usage{
... \%notin\% NA
}
\description{
negates the \%in\% infix op, looks for things that are not in your list
}
\examples{
c(1,2,3) \%notin\% c(3,4,5)
}
