% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_3kpk_scores.R
\name{prep_3kpk_scores}
\alias{prep_3kpk_scores}
\title{Get 3KPK 1395 RFP ReadScores ready prepared for push into fse files}
\usage{
prep_3kpk_scores(df)
}
\arguments{
\item{df}{}
}
\value{
a gropued list broken out by evaluator and proposal id
}
\description{
Get 3KPK 1395 RFP ReadScores ready prepared for push into fse files
}
