% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_charter_scores.R
\name{prep_charter_scores}
\alias{prep_charter_scores}
\title{Prep charter scores for passport export}
\usage{
prep_charter_scores(df)
}
\arguments{
\item{df}{charter interview scores to export}
}
\value{

}
\description{
Prep charter scores for passport export
}
