% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores.R
\name{prep_scores}
\alias{prep_scores}
\title{Prep scores for score export}
\usage{
prep_scores(df_read, df_interview)
}
\arguments{
\item{df_interview}{}
}
\value{
3 lists of the scores to export over
}
\description{
Prep scores for score export
}
