% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_installed_packages.R
\name{save_installed_packages}
\alias{save_installed_packages}
\title{save_installed_packages}
\usage{
save_installed_packages(path_to_save = NULL)
}
\arguments{
\item{path_to_save}{character vector, needs to be a csv file path to save to}
}
\value{
no return, simply exports the file as a csv
}
\description{
save_installed_packages
}
\examples{
x <- tempdir()
save_installed_packages(path_to_save = paste0(x, "my_pacakges.csv"))
}
