% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores.R
\name{prep_scores}
\alias{prep_scores}
\title{Get 3KPK 1395 RFP ReadScores ready prepared for push into fse files}
\usage{
prep_scores(df_read, df_interview)
}
\arguments{
\item{df_read}{ReadScore export from VSET backend}

\item{df_interview}{Interview score exports from vset backend}
}
\value{
a grouped list broken out by evaluator and proposal id
}
\description{
Get 3KPK 1395 RFP ReadScores ready prepared for push into fse files
}
